Consolidate host builders and make static files work in WPF
