+++ src/Microsoft.MobileBlazorBindings.Hosting/Microsoft.MobileBlazorBindings.Hosting.csproj
@@ -12,6 +12,7 @@
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.Components.Web" Version="3.1.10" />
     <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.10" />
+    <PackageReference Include="Microsoft.Extensions.Logging" Version="3.1.10" />
   </ItemGroup>
 
   <ItemGroup>
