+++ src/Microsoft.MobileBlazorBindings/MobileBlazorBindingsHost.cs
@@ -1,13 +1,9 @@
 ﻿// Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System.IO;
-using System.Reflection;
-using System.Runtime.InteropServices;
+using Microsoft.MobileBlazorBindings.Hosting;
 
 namespace Microsoft.MobileBlazorBindings
 {
@@ -15,49 +11,7 @@ namespace Microsoft.MobileBlazorBindings
     {
         public static IHostBuilder CreateDefaultBuilder(string[] args = null)
         {
-            // Inspired by Microsoft.Extensions.Hosting.Host, which can be seen here:
-            // https://github.com/dotnet/runtime/blob/master/src/libraries/Microsoft.Extensions.Hosting/src/Host.cs
-            // But slightly modified to work on all of Android, iOS, and UWP.
-
-            var builder = new HostBuilder();
-
-            builder.UseContentRoot(Directory.GetCurrentDirectory());
-            builder.UseWebRoot("wwwroot");
-            builder.ConfigureAppConfiguration((hostingContext, config) =>
-            {
-                var env = hostingContext.HostingEnvironment;
-
-                config.AddJsonFile("appsettings.json", optional: true, reloadOnChange: RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-                      .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true, reloadOnChange: RuntimeInformation.IsOSPlatform(OSPlatform.Windows));
-
-                if (env.IsDevelopment())
-                {
-                    var appAssembly = Assembly.Load(new AssemblyName(env.ApplicationName));
-                    if (appAssembly != null)
-                    {
-                        config.AddUserSecrets(appAssembly, optional: true);
-                    }
-                }
-
-                config.AddEnvironmentVariables();
-
-                if (args != null)
-                {
-                    config.AddCommandLine(args);
-                }
-            });
-            builder.ConfigureLogging((hostingContext, logging) =>
-            {
-                logging.AddConsole(configure => configure.DisableColors = true);
-                logging.AddDebug();
-                logging.AddEventSourceLogger();
-            })
-            .UseDefaultServiceProvider((context, options) =>
-            {
-                var isDevelopment = context.HostingEnvironment.IsDevelopment();
-                options.ValidateScopes = isDevelopment;
-                options.ValidateOnBuild = isDevelopment;
-            });
+            var builder = BlazorWebHost.CreateDefaultBuilder(args);
 
             EnableStyleSheetSupport();
